Begin rewrite. - Namespace changed to JaxkDev/KOTH - PSR-4 now used. - File header reduced, end year removed, discord tag corrected. - Removed YAML - Added PHPStan/PMMP to composer dev requirements.
