+++ composer.json
@@ -1,0 +1,21 @@
+{
+    "license": "GPL-3.0",
+    "authors": [
+        {
+            "name": "JaxkDev",
+            "email": "JaxkDev@gmail.com"
+        }
+    ],
+    "autoload": {
+        "psr-4": {
+            "JaxkDev\\KOTH\\": "src"
+        }
+    },
+    "require": {
+        "php": "^8.0.0"
+    },
+    "require-dev": {
+        "phpstan/phpstan": "^1.7",
+        "pocketmine/pocketmine-mp": "^4"
+    }
+}
