+++ phpstan.neon
@@ -1,10 +1,7 @@
 parameters:
-	level: 5
-	paths:
-		- /source/src
-	autoload_files:
-		- phar:///pocketmine/PocketMine-MP.phar/vendor/autoload.php
-	autoload_directories:
-		- /source/src
-		- /deps
-	inferPrivatePropertyTypeFromConstructor: true
+    level: max
+    paths:
+        - src
+    excludePaths:
+        analyse:
+            - vendor
