+++ resources/config.yml
@@ -5,10 +5,7 @@ provider: "sqlite3"
 
 debug: false
 
-show_updates: true
 check_updates: true
-download_updates: false
-update_check_url: "https://raw.githubusercontent.com/JaxkDev/koth/master/updates.json"
 
 block_commands: true
 block_messages: true
