+++ src/Tasks/Prestart.php
@@ -1,41 +1,33 @@
 <?php
 /*
-*    /$$   /$$  /$$$$$$  /$$$$$$$$ /$$   /$$
-*   | $$  /$$/ /$$__  $$|__  $$__/| $$  | $$
-*   | $$ /$$/ | $$  \ $$   | $$   | $$  | $$
-*   | $$$$$/  | $$  | $$   | $$   | $$$$$$$$
-*   | $$  $$  | $$  | $$   | $$   | $$__  $$
-*   | $$\  $$ | $$  | $$   | $$   | $$  | $$
-*   | $$ \  $$|  $$$$$$/   | $$   | $$  | $$
-*   |__/  \__/ \______/    |__/   |__/  |__/
-*
-*   Copyright (C) 2019-2020 JaxkDev
-*
-*   This program is free software: you can redistribute it and/or modify
-*   it under the terms of the GNU General Public License as published by
-*   the Free Software Foundation, either version 3 of the License, or
-*   any later version.
-*
-*   This program is distributed in the hope that it will be useful,
-*   but WITHOUT ANY WARRANTY; without even the implied warranty of
-*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*   GNU General Public License for more details.
-*
-*   You should have received a copy of the GNU General Public License
-*   along with this program.  If not, see <https://www.gnu.org/licenses/>.
-*
-*   Twitter :: @JaxkDev
-*   Discord :: JaxkDev#8860
-*   Email   :: JaxkDev@gmail.com
-*/
+ *   KOTH, A pocketmine-MP Mini-game
+ *
+ *   Copyright (C) 2019-present JaxkDev
+ *
+ *   This program is free software: you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation, either version 3 of the License, or
+ *   any later version.
+ *
+ *   This program is distributed in the hope that it will be useful,
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *   GNU General Public License for more details.
+ *
+ *   You should have received a copy of the GNU General Public License
+ *   along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ *   Twitter :: @JaxkDev
+ *   Discord :: JaxkDev#2698
+ *   Email   :: JaxkDev@gmail.com
+ */
 
-declare(strict_types=1);
-namespace Jackthehack21\KOTH\Tasks;
+namespace JaxkDev\KOTH\Tasks;
 
 use pocketmine\scheduler\Task;
 
-use Jackthehack21\KOTH\Main;
-use Jackthehack21\KOTH\Arena;
+use JaxkDev\KOTH\Main;
+use JaxkDev\KOTH\Arena;
 
 class Prestart extends Task{
